File can now be persisted
Fixed problem with the helper module. Needed to use the correct Electron syntax (different from regualr js), not to use toString() on the date methods (difference bettwen js and java) and correct and ommision of brackets () after calling the dateStr helper method in main.js

On this branch I'll keep the results of the failed attemps - good lesson to learn from with the filenaming.