return to hashing email before storing
Just thought that maybe it really is a better idea to do this (since it's possible in this case as we never need the email again except to check that email for reset code matches stored hash). If someone happens to access that app, they will not get the assosicated email from it.